update the exercise task

+++ db/migrate/20190301013336_create_exercises.rb
@@ -5,7 +5,7 @@ class CreateExercises < ActiveRecord::Migration[5.2]
       t.string :name, null: false
       t.string :exercise_type, null: false
       t.string :location, null: false
-      t.integer :distance, null: false
+      t.float :distance, null: false
       t.float :start_lnt, null: false
       t.float :start_lat, null: false
       t.float :end_lnt, null: false
